product menu ( items bought together)
The algorithm we use here is the association rules mining algorithm, or more properly called, marked basket analysis. Association rules mining are put rules that specify the dependency of one item onto another. It would be like if you bought a laptop, then carry cases are bought together 60% of the time.